fix(main): レビューコメントを反映
Uvicornのライフサイクルイベントを利用したグレースフルシャットダウンを実装しました。
